requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
arrow==1.4.0
attrs==25.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
codecarbon==3.2.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dasheng==0.0.8
einops==0.8.1
fief-client==0.20.0
filelock==3.19.1
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
jwcrypto==1.5.6
kiwisolver==1.4.9
lightning==2.5.3
lightning-utilities==0.15.2
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
numpy==2.3.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
nvtx==0.2.14
omegaconf==2.3.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.0
psutil==7.2.1
py-cpuinfo==9.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-ignite==0.5.2
pytorch-lightning==2.5.3
pytz==2025.2
PyYAML==6.0.2
questionary==2.1.1
RapidFuzz==3.14.3
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
torch==2.8.0
torchaudio==2.8.0
torchcodec==0.6.0
torchmetrics==1.8.1
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
wcwidth==0.2.14
Werkzeug==3.1.3
yarl==1.20.1
yaspin==3.4.0
